feat(openshift): add support for OpenShift DeploymentConfig (#95)
Add Locator, Mutator, and Revertor for handling the OpenShift DeploymentConfig kind.

The Controller will select the first Locator with a positive response and ignore the rest. A DeploymentConfig is cloned similarly as k8s core Deployment kinds.

Add option to create Deployment or DeploymentConfig as a vessel for deployment in the Test Scenario Generator. 

Requires Telepresence v 0.100, which contain a fix for the DeploymentConfig kind (#96).

Fixes #50