feat(cli): add namespace bound install flag (#217)
A new flag --local has been added to the "install-operator" command that only gives the operator namespace bound permissions and the operator will only watch the namespace it is installed in.

No cluster-wide auth required to run, but still require cluster-admin to install CRDs. Multiple 'local' Operators can operate on the same cluster, but not combined with a cluster-wide operator. 

E2E test suite changed to use the new local install option.

Related to #214