feat: enables defining strategy templates in custom folder (#602)
Before this change strategy templates used to patch k8s objects were defined in the code. This allows to define them in an arbitrary folder which should be then exposed as environment variable TEMPLATE_PATH. This folder should have .tpl and .var files named after the strategy. Former defined operations whereas the latter have a list of variables that can be set to empty and serves mainly as documentation for what can be passed into the template or contain a default value.

Changes
* loading templates from filesystem or assets linked by go-bindata
* handling external template folder by using TEMPLATE_PATH
* minor code improvements as a result of this impl changes

Fixes #203

Co-authored-by: Aslak Knutsen <aslak@4fs.no>